Client: add optional SOCKS5 proxy
Adds a new option to ServerConfig, proxy_address.
If proxy_address is defined, it will be used to
add a Socks5ProxyHandler to the netty pipeline.
This also fixes several NullPointExceptions and
incorrect adornments on Client.handleDisconnect.

By default, the value of proxy_address (and
comms.address) is 'none!' to prevent unintentional
connections to a remote server of any sort.

It is recommended to copy+paste when configuring
both address and proxy_address, since the
configuration panel will update the configuration
and reconnect to it's value every time it's
updated, which includes while your typing (Thanks
gjum).

Client: remove an annoyingly loud debug log

This log constantly fires when a valid server is
configured, but the proxy is not. Just going to
remove it.