decorations: display decorations on top when legal
After having a conversation with Tritonio about
the legality of the 'glow' shader provided by
minecraft, I am going to assume that it's not
legal on CivRealms.

In order to make it civ legal, the 'glow'
decorator has been removed from Axon and the other
decorators have been made to render without depth
where legal.