synchronization: make the server work, fix client.
Takes the server module from Synapse and makes
it actually work. This also changes how the client
synchronizes with the server. Instead of sending
detailed information about itself, it will only
send the standard PlayerState.

Why? Because, it doesn't work, and I can't be
bothered to find out why. It works just as well
either way.

In order to use the server, run it once so it
creates the files it needs. Then open uuids.tsv
and username/uuid pairs into the file.

In order to whitelist them, you put their username
into users.tsv. It's stupid, you can complain to
gjum.