decorators: fix depth/lighting issue with tile ents
Changes the event that triggers the decorator
rendering from onPostEntityRender to onPostRender.

Likely culprit was that later rendering calls were
occluding the decorator. By running everything last
we fix this issue.

The downside is, depth does not work *at all*, so
we have to fully disable all decorators when a
player is invisible or crouching in order to be
civ-legal.